857: Move old calendar export controller code into legacy controller
