+++ app/views/calendars/exports/index.html.erb
@@ -67,6 +67,6 @@
 
 <div class="footnote">
   If you accidentally share your personalized calendar links and want to make them private again,
-  <%= link_to("click here to reset your secret token", reset_token_calendars_exports_path, method: :put) %>.
+  <%= link_to("click here to reset your secret token", calendars_exports_reset_token_path, method: :put) %>.
   Note: This will make the old links inoperative, so you will need to update your calendar app.
 </div>
